feat(runner): add WallTimeExecutor
