+++ src/run/runner/helpers/mod.rs
@@ -1,1 +1,4 @@
+pub mod env;
+pub mod get_bench_command;
 pub mod profile_folder;
+pub mod run_command_with_log_pipe;
