+++ src/run/runner/interfaces.rs
@@ -9,4 +9,15 @@ pub struct RunData {
 #[serde(rename_all = "lowercase")]
 pub enum ExecutorName {
     Valgrind,
+    WallTime,
+}
+
+#[allow(clippy::to_string_trait_impl)]
+impl ToString for ExecutorName {
+    fn to_string(&self) -> String {
+        match self {
+            ExecutorName::Valgrind => "valgrind".to_string(),
+            ExecutorName::WallTime => "walltime".to_string(),
+        }
+    }
 }
