+++ src/run/runner/wall_time/executor.rs
@@ -1,0 +1,75 @@
+use crate::prelude::*;
+
+use crate::run::instruments::mongo_tracer::MongoTracer;
+use crate::run::runner::executor::Executor;
+use crate::run::runner::helpers::env::BASE_INJECTED_ENV;
+use crate::run::runner::helpers::get_bench_command::get_bench_command;
+use crate::run::runner::helpers::run_command_with_log_pipe::run_command_with_log_pipe;
+use crate::run::runner::{ExecutorName, RunData};
+use crate::run::{check_system::SystemInfo, config::Config};
+use async_trait::async_trait;
+use std::fs::canonicalize;
+use std::process::Command;
+
+const WALL_TIME_EXECUTION_TARGET: &str = "walltime::execution";
+pub const WALL_TIME_RUNNER_MODE: &str = "walltime";
+
+pub struct WallTimeExecutor;
+
+#[async_trait(?Send)]
+impl Executor for WallTimeExecutor {
+    fn name(&self) -> ExecutorName {
+        ExecutorName::WallTime
+    }
+
+    async fn setup(
+        &self,
+        _config: &Config,
+        _system_info: &SystemInfo,
+        _run_data: &RunData,
+    ) -> Result<()> {
+        Ok(())
+    }
+
+    async fn run(
+        &self,
+        config: &Config,
+        _system_info: &SystemInfo,
+        run_data: &RunData,
+        _mongo_tracer: &Option<MongoTracer>,
+    ) -> Result<()> {
+        let mut cmd = Command::new("sh");
+
+        cmd.env("CODSPEED_RUNNER_MODE", &self.name().to_string())
+            .env(
+                "CODSPEED_PROFILE_FOLDER",
+                run_data.profile_folder.to_str().unwrap(),
+            )
+            .envs(BASE_INJECTED_ENV.iter());
+
+        if let Some(cwd) = &config.working_directory {
+            let abs_cwd = canonicalize(cwd)?;
+            cmd.current_dir(abs_cwd);
+        }
+
+        cmd.args(["-c", get_bench_command(config)?.as_str()]);
+
+        debug!("cmd: {:?}", cmd);
+        let status = run_command_with_log_pipe(cmd, WALL_TIME_EXECUTION_TARGET)
+            .map_err(|e| anyhow!("failed to execute the benchmark process. {}", e))?;
+        if !status.success() {
+            bail!("failed to execute the benchmark process");
+        }
+
+        Ok(())
+    }
+
+    async fn teardown(
+        &self,
+        _config: &Config,
+        _system_info: &SystemInfo,
+        _run_data: &RunData,
+    ) -> Result<()> {
+        Ok(())
+    }
+}
